Refactored the DV360 run query operator following the Concerto guidelines.
